MailVerificationManager, FriendsDAO, cambios BD

+++ DataBaseManager/DAO/FriendsDAO.cs
@@ -19,34 +19,64 @@ namespace DataBaseManager.DAO {
 
                     db.Amistad.Add(newFriendship);
                     db.SaveChanges();
-                    operationStatus = Constants.SUCESS;
+                    operationStatus = Constants.SUCCSESS;
                 }
-            } catch (EntityException entityException) {
+            }
+            catch (EntityException entityException) {
                 Console.WriteLine("Error trying to register the friendship {0}", entityException.Message);
             }
             return operationStatus;
         }
 
+        //Not useful anymore?
         public int deleteFriendshipDAO(int idProfile1, int idProfile2) {
             int operationStatus = Constants.FAILED;
             try {
                 using (tripasEntities db = new tripasEntities()) {
-                    Amistad friendshipToDelete = new Amistad() {
-                        idJugadorUno = idProfile1,
-                        idJugadorDos = idProfile2,
-                    };
+                    var friendshipToDelete = db.Amistad.FirstOrDefault(a =>
+                    a.idJugadorUno == idProfile1 && a.idJugadorDos == idProfile2);
 
-                    db.Amistad.Remove(friendshipToDelete);
-                    db.SaveChanges();
-                    operationStatus = Constants.SUCESS;
+                    if (friendshipToDelete != null) {
+                        // Si se encontró la amistad, eliminarla
+                        db.Amistad.Remove(friendshipToDelete);
+                        db.SaveChanges();
+                        operationStatus = Constants.SUCCSESS;
+                    }
                 }
-            } catch (EntityException entityException) {
+            }
+            catch (EntityException entityException) {
                 Console.WriteLine(entityException.Message);
             }
             return operationStatus;
         }
 
-        public List<Perfil> getFriendsDAO (int idProfile) {
+        public int deleteFriendsDAO(string userName1, string userName2) {
+            int operationStatus = Constants.FAILED;
+            try {
+                using (tripasEntities db = new tripasEntities()) {
+                    var profile1 = db.Perfil.FirstOrDefault(profile => profile.nombre == userName1);
+                    var profile2 = db.Perfil.FirstOrDefault(profile => profile.nombre == userName2);
+                    if (profile1 != null && profile2 != null) {
+                        var friendshipToDelete = db.Amistad.FirstOrDefault(perfil =>
+                         perfil.idJugadorUno == profile1.idPerfil && perfil.idJugadorDos == profile2.idPerfil);
+
+                        if (friendshipToDelete != null) {
+                            db.Amistad.Remove(friendshipToDelete);
+                            db.SaveChanges();
+                            operationStatus = Constants.SUCCSESS;
+                        }
+                    }
+                }
+            }
+            catch (EntityException entityException) {
+                Console.WriteLine("Unable to delete friendship " + entityException.Message);
+            }
+            return operationStatus;
+        }
+
+
+
+        public List<Perfil> getFriendsDAO(int idProfile) {
             List<Perfil> friendList = new List<Perfil>();
             try {
                 using (tripasEntities db = new tripasEntities()) {
@@ -56,7 +86,8 @@ namespace DataBaseManager.DAO {
                                    select profile).ToList();
                     friendList = friends;
                 }
-            } catch (EntityException entityException) {
+            }
+            catch (EntityException entityException) {
                 Console.WriteLine("Error trying to retrieve the friend list {0}", entityException.Message);
             }
             return friendList;
